Fix db initialisation
