+++ codecov.yml
@@ -6,6 +6,3 @@ coverage:
     patch:
       default:
         informational: true
-
-ignore:
-  - "backend/bracket/utils/db_init.py"
